+++ daemon/state.go
@@ -5,8 +5,8 @@ import (
 	"sync"
 	"time"
 
-	derr "github.com/docker/docker/api/errors"
 	"github.com/docker/docker/daemon/execdriver"
+	derr "github.com/docker/docker/errors"
 	"github.com/docker/docker/pkg/units"
 )
