+++ errors/daemon.go
